pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imap-mcp"
version = "0.1.0"
description = "IMAP Model Context Protocol server for interactive email processing"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Nikita"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "mcp>=1.26.0",
    "imapclient>=3.1.0",
    "email-validator>=2.3.0",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.1",
    "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
]

[project.urls]
Homepage = "https://github.com/ivni/imap-mcp"
Issues = "https://github.com/ivni/imap-mcp/issues"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["imap_mcp"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "-v"
log_cli = true
log_cli_level = "ERROR"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
markers = [
    "unit: Unit tests (default)",
    "integration: Integration tests requiring actual services",
    "slow: Tests that take a long time to run",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
